4.23 bloglist step11

+++ controllers/blog.js
@@ -1,8 +1,10 @@
+// const jwt = require('jsonwebtoken')
 const blogRouter = require('express').Router()
+const middleware = require('../utils/middleware')
 const Blog = require('../models/blog')
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
@@ -17,28 +19,51 @@ blogRouter.get('/:id', async (request, response) => {
   }
 })
 
-blogRouter.post('/', async (request, response) => {
-  const body = request.body
+blogRouter.post('/', middleware.userExtractor, async (request, response) => {
+  const { url, title, author, likes } = request.body
 
-  if (!body.title || !body.url) {
+  const user = request.user
+
+  if (!(title || url)) {
     return response.status(400).json({ error: 'title or url missing' })
   }
 
   const blog = new Blog({
-    title: body.title,
-    author: body.author,
-    url: body.url,
-    likes: body.likes || 0,
+    url: url,
+    title: title,
+    author: author,
+    user: user.id,
+    likes: likes || 0,
   })
 
   const savedBlog = await blog.save()
+  user.blogs = user.blogs.concat(savedBlog._id)
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
-blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-})
+blogRouter.delete(
+  '/:id',
+  middleware.userExtractor,
+  async (request, response) => {
+    const user = request.user
+    const blog = await Blog.findById(request.params.id)
+
+    if (!blog) {
+      return response.status(404).end()
+    }
+
+    if (user.id.toString() !== blog.user.toString()) {
+      return response.status(401).json({
+        error: 'this blog can be deleted only by the user who created it.',
+      })
+    }
+
+    await Blog.findByIdAndDelete(request.params.id)
+    response.status(204).end()
+  }
+)
 
 blogRouter.put('/:id', async (request, response) => {
   const { title, author, url, likes } = request.body
